requirements.txt
pytest
pyyaml
boto3
requests
Jinja2>=3.1.0
